jobpilot-ai/requirements.txt
# ============================================================
# JobPilot AI - Requirements
# ============================================================

# --- Core Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.1
python-multipart==0.0.9

# --- LLM (Provider-Agnostic) ---
litellm==1.40.0
openai==1.35.0
anthropic==0.30.0
tiktoken==0.7.0

# --- Browser Automation (AI-native agent) ---
playwright==1.44.0
browser-use==0.1.29

# --- Resume Parsing ---
PyMuPDF==1.24.5          # PDF parsing
pdfplumber==0.11.0        # PDF table extraction
python-docx==1.1.2        # DOCX resume parsing

# --- ML / NLP ---
sentence-transformers==3.0.1     # Semantic embeddings
scikit-learn==1.5.0              # Cosine similarity, ML utils
numpy==1.26.4
pandas==2.2.2

# --- Vector Store (local) ---
chromadb==0.5.3

# --- Database ---
sqlalchemy==2.0.31
alembic==1.13.1
aiosqlite==0.20.0

# --- Async ---
httpx==0.27.0
aiofiles==23.2.1

# --- MCP Server ---
fastmcp==0.4.1
mcp==1.0.0

# --- UI ---
streamlit==1.36.0
plotly==5.22.0
streamlit-extras==0.4.3

# --- Analytics ---
matplotlib==3.9.0
seaborn==0.13.2

# --- Utilities ---
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1
typer==0.12.3
jinja2==3.1.4

# --- Testing ---
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
